signup/login secured
while signing up or logging in sending user data to server using encryption & added the packages necessary.
