+++ .gitignore
@@ -17,7 +17,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
-config.env
+*.env
 
 npm-debug.log*
 yarn-debug.log*
